+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/index.ts
@@ -5,6 +5,6 @@ export * from './select-flag-location/select-flag-location.component';
 export * from './add-comments/add-comments.component';
 export * from './manage-case-flags/manage-case-flags.component';
 export * from './update-flag/update-flag.component';
-export * from './update-flag/update-flag-welsh-translation-form/update-flag-welsh-translation-form.component';
+export * from './update-flag/update-flag-add-translation-form/update-flag-add-translation-form.component';
 export * from './case-flag-summary-list/case-flag-summary-list.component';
 export * from './confirm-flag-status/confirm-flag-status.component';