+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/index.ts
@@ -1,5 +1,6 @@
 export * from './add-comments-error-message.enum';
 export * from './add-comments-step.enum';
+export * from './case-flag-form-fields.enum';
 export * from './case-flag-status.enum';
 export * from './case-flag-summary-list-display-mode.enum';
 export * from './case-flag-wizard-step-title.enum';
@@ -10,6 +11,8 @@ export * from './search-language-interpreter-step.enum';
 export * from './select-flag-error-message.enum';
 export * from './select-flag-location-error-message.enum';
 export * from './select-flag-type-error-message.enum';
-export * from './write-case-flag-field.enum';
+export * from './update-flag-add-translation-error-message.enum';
+export * from './update-flag-add-translation-step.enum';
 export * from './update-flag-error-message.enum';
 export * from './update-flag-step.enum';
+export * from './write-case-flag-field.enum';